+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Retrieving BIM360 issue container IDs
+
 ## [8.0.0] - 2019-11-02
 
 ### Added